Update 20250112 - Base Class
Added new base classes which will serve as a superclass of all entity model class.

[CHANGELOG]
🟢 Added Base Model class which will handle all the common rendering logic.
🟢 Added Base Projectile Model class which will handle all projectile-based rendering logics.
🟢 Added annotation to the animation path parameter for the Base Turret Model for parity purposes and for potential use-case in the future.
🟡 Updated the Ballista Arrow model class to now extend the new Base Projectile class, reducing its code significantly.
🟡 Updated the Cannonball model class to now extend the new Base Projectile class, reducing its code significantly.
